pyproject.toml
[project]
name = "bethesda-shelter-agent"
version = "0.1.0"
description = "AI Voice Agent for Bethesda Mission Men's Shelter - 108 bed management system"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Web Framework
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    
    # Telephony
    "twilio>=8.10.0",
    
    # AI/ML
    "openai>=1.3.0",
    "pinecone>=7.0.0",
    
    # Database
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.12.0",
    
    # Background Jobs
    "celery>=5.3.0",
    "redis>=5.0.0",
    
    # Utilities
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.25.0",
    
    # Security
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.6",
    "greenlet>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.11.0",
    "ruff>=0.1.6",
    "mypy>=1.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.ruff]
line-length = 88
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.11"
strict = true
